Add files via upload
This code is about a collection of data from a company wishing to process its customer 
records for business purposes (acw_user_data.csv). The existing systems in-place at the 
company only export to a CSV file, and this is not in an appropriate format for analysis. i i i will like to assume that i have been given the task of preparing this data for further analyses by my colleagues within the company, including representation changes, filtering, and deriving some new attributes / metrics for them.